+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Released on Tuesday, May 31 2022.
 - Dropped .NET Framework 4.6
 - Updated to use AngleSharp 0.17
 - Fixed casing issue with color, timing, and gradient functions (#109)
+- Fixed shorthand properties using `inherit` being omitted (#100)
 
 # 0.16.4